refactor: implement requested changes by Mike Lankamp
